Added farming capacity limit

+++ cardinal/component/farming.go
@@ -2,7 +2,7 @@ package component
 
 type Farming struct {
 	Type     ResourceType `json:"type"`
-	Speed    float32      `json:"speed"`
+	Speed    float64      `json:"speed"`
 	Capacity int          `json:"capacity,omitempty"`
 }
